+++ digiplan/static/js/sliders.js
@@ -1,7 +1,5 @@
 // Variables
 const SETTINGS_PARAMETERS = JSON.parse(document.getElementById("settings_parameters").textContent);
-const SETTINGS_DEPENDENCY_MAP = JSON.parse(document.getElementById("settings_dependency_map").textContent);
-const DEPENDENCY_PARAMETERS = JSON.parse(document.getElementById("dependency_parameters").textContent);
 const panelContainer = document.getElementById("js-panel-container");
 const panelSliders = document.querySelectorAll(".js-slider.js-slider-panel");
 const powerPanelSliders = document.querySelectorAll(".js-slider.js-slider-panel.js-power-mix");
@@ -23,54 +21,16 @@ const potentialWindLayers = [
   "potentialarea_wind_stp_2027_search_area_open_area",
   "potentialarea_wind_stp_2027_vr"
 ];
-const potentialWindSwitches = document.querySelectorAll("#id_s_w_3, #id_s_w_4, #id_s_w_4_1, #id_s_w_4_2, #id_s_w_5, #id_s_w_5_1, #id_s_w_5_2");
 
-const sectorSlider = document.querySelectorAll("#id_s_v_3, #id_s_v_4, #id_s_v_5, #id_w_d_wp_3, #id_w_d_wp_4, #id_w_d_wp_5, #id_w_v_3, #id_w_v_4, #id_w_v_5");
-
-const sliderDependencies = {
-  "id_v_iv_1": "id_v_iv_3"
-};
 
 // Setup
 
-// Order matters. Start with the most specific, and end with most general sliders.
-Array.from(Object.keys(SETTINGS_DEPENDENCY_MAP)).forEach(dependent_name => {
-  const dependency_names = SETTINGS_DEPENDENCY_MAP[dependent_name];
-  Array.from(dependency_names).forEach(dependency_name => {
-    $("#id_" + dependency_name).ionRangeSlider({
-        onChange: function (data) {
-          const msg = eventTopics.DEPENDENCY_PANEL_SLIDER_CHANGE;
-          PubSub.publish(msg, {
-            dependent: dependent_name,
-            dependency: dependency_name,
-            data
-          });
-        }
-      }
-    );
-  });
-});
 $(".js-slider.js-slider-panel.js-power-mix").ionRangeSlider({
     onChange: function (data) {
       PubSub.publish(eventTopics.POWER_PANEL_SLIDER_CHANGE, data);
     }
   }
 );
-$(potentialWindSwitches).on("change", function () {
-  PubSub.publish(eventTopics.WIND_CONTROL_ACTIVATED);
-});
-$(".js-slider.js-slider-panel").ionRangeSlider({
-    onChange: function (data) {
-      PubSub.publish(eventTopics.PANEL_SLIDER_CHANGE, data);
-    }
-  }
-);
-$(sectorSlider).ionRangeSlider({
-    onChange: function (data) {
-      calculate_slider_value(data);
-    }
-  }
-);
 
 $(".js-slider.js-slider-panel").ionRangeSlider({
     onChange: function (data) {
@@ -78,44 +38,10 @@ $(".js-slider.js-slider-panel").ionRangeSlider({
     }
   }
 );
-$(".form-check-input").on(
-    'click', function (data) {
-      toggleFormFields(data.target.id);
-    }
-);
-$("#id_s_w_5_1").ionRangeSlider({
+
+$(".js-slider.js-slider-detail-panel").ionRangeSlider({
     onChange: function (data) {
-      calculate_max_wind();
-    }
-  }
-);
-$("#id_s_w_5_2").ionRangeSlider({
-    onChange: function (data) {
-      calculate_max_wind();
-    }
-  }
-);
-$("#id_s_pv_ff_3").ionRangeSlider({
-    onChange: function (data) {
-      calculate_max_pv_ff();
-    }
-  }
-);
-$("#id_s_pv_ff_4").ionRangeSlider({
-    onChange: function (data) {
-      calculate_max_pv_ff();
-    }
-  }
-);
-$("#id_s_pv_d_3").ionRangeSlider({
-    onChange: function (data) {
-      calculate_max_pv_d();
-    }
-  }
-);
-$("#id_v_iv_3").ionRangeSlider({
-    onChange: function (data) {
-      $(`#id_v_iv_1`).data("ionRangeSlider").update({from:data.from});
+      PubSub.publish(eventTopics.DETAIL_PANEL_SLIDER_CHANGE, data);
     }
   }
 );
@@ -140,38 +66,25 @@ subscribeToEvents(
   [eventTopics.STATES_INITIALIZED, eventTopics.POWER_PANEL_SLIDER_CHANGE],
   createPercentagesOfPowerSources
 );
-subscribeToEvents(
-  [eventTopics.POWER_PANEL_SLIDER_CHANGE, eventTopics.PANEL_SLIDER_CHANGE],
-  showActivePanelSliderOnPanelSliderChange
-);
-subscribeToEvents(
-  [eventTopics.POWER_PANEL_SLIDER_CHANGE, eventTopics.PANEL_SLIDER_CHANGE],
-  hidePotentialLayers
-);
-subscribeToEvents(
-  [eventTopics.POWER_PANEL_SLIDER_CHANGE, eventTopics.PANEL_SLIDER_CHANGE],
-  checkMainPanelSlider
-);
+PubSub.subscribe(eventTopics.PANEL_SLIDER_CHANGE, showActivePanelSliderOnPanelSliderChange);
+PubSub.subscribe(eventTopics.PANEL_SLIDER_CHANGE, hidePotentialLayers);
+PubSub.subscribe(eventTopics.PANEL_SLIDER_CHANGE, adaptDetailSliders);
+PubSub.subscribe(eventTopics.DETAIL_PANEL_SLIDER_CHANGE, adaptMainSliders);
 PubSub.subscribe(eventTopics.MORE_LABEL_CLICK, showOrHideSidepanelsOnMoreLabelClick);
 PubSub.subscribe(eventTopics.MORE_LABEL_CLICK, showOrHidePotentialLayersOnMoreLabelClick);
-PubSub.subscribe(eventTopics.DEPENDENCY_PANEL_SLIDER_CHANGE, (msg, payload) => {
-  const {dependent, dependency, data} = payload;
-  const value = DEPENDENCY_PARAMETERS[dependency][dependent][data.from];
-  const dependentDataElement = $("#id_" + dependent).data("ionRangeSlider");
-  dependentDataElement.update({max: value});
-});
 PubSub.subscribe(eventTopics.PV_CONTROL_ACTIVATED, showPVLayers);
 PubSub.subscribe(eventTopics.WIND_CONTROL_ACTIVATED, showWindLayers);
 
 
 // Subscriber Functions
 
-function checkMainPanelSlider(msg, data) {
-  if (sliderDependencies.hasOwnProperty(data.input[0].id)) {
-    let target = sliderDependencies[data.input[0].id];
-    $('#' + target).data("ionRangeSlider").update({from:data.from});
-  }
-    if (data.input[0].id === "id_s_v_1") {
+/**
+ * Adapt detail sliders depending on related main sliders
+ * @param {string} msg Publisher message
+ * @param {object} data Data from changed ionrangeslider
+ */
+function adaptDetailSliders(msg, data) {
+  if (data.input[0].id === "id_s_v_1") {
     $(`#id_s_v_3`).data("ionRangeSlider").update({from:data.from});
     $(`#id_s_v_4`).data("ionRangeSlider").update({from:data.from});
     $(`#id_s_v_5`).data("ionRangeSlider").update({from:data.from});
@@ -186,6 +99,56 @@ function checkMainPanelSlider(msg, data) {
     $(`#id_w_v_4`).data("ionRangeSlider").update({from:data.from});
     $(`#id_w_v_5`).data("ionRangeSlider").update({from:data.from});
   }
+  return logMessage(msg);
+}
+
+/**
+ * Adapt main slider depending on related detail sliders
+ * @param {string} msg Publisher message
+ * @param {object} data Data from changed ionrangeslider
+ */
+function adaptMainSliders(msg, data) {
+  const slider_id = data.input[0].id;
+  if (slider_id === "id_s_w_6" || slider_id === "id_s_w_7") {
+    calculate_max_wind();
+  }
+  if (slider_id === "id_s_pv_d_3") {
+    calculate_max_pv_d();
+  }
+  if (slider_id === "id_s_pv_ff_3" || slider_id === "id_s_pv_ff_4") {
+    calculate_max_pv_ff();
+  }
+  if (slider_id === "id_s_v_3" || slider_id === "id_s_v_4" || slider_id === "id_s_v_5") {
+    let factor_hh = $("#id_s_v_3").data("ionRangeSlider").result.from;
+    let factor_ind = $("#id_s_v_5").data("ionRangeSlider").result.from;
+    let factor_cts = $("#id_s_v_4").data("ionRangeSlider").result.from;
+    let demand_hh = store.cold.slider_per_sector.s_v_1.hh;
+    let demand_ind = store.cold.slider_per_sector.s_v_1.ind;
+    let demand_cts = store.cold.slider_per_sector.s_v_1.cts;
+    let new_val = (factor_hh * demand_hh + factor_ind * demand_ind + factor_cts * demand_cts) / (demand_hh + demand_ind + demand_cts);
+    $(`#id_s_v_1`).data("ionRangeSlider").update({from:new_val});
+  }
+  if (slider_id === "id_w_d_wp_3" || slider_id === "id_w_d_wp_4" || slider_id === "id_w_d_wp_5") {
+    let factor_hh = $("#id_w_d_wp_3").data("ionRangeSlider").result.from;
+    let factor_ind = $("#id_w_d_wp_4").data("ionRangeSlider").result.from;
+    let factor_cts = $("#id_w_d_wp_5").data("ionRangeSlider").result.from;
+    let demand_hh = store.cold.slider_per_sector.w_d_wp_1.hh;
+    let demand_ind = store.cold.slider_per_sector.w_d_wp_1.ind;
+    let demand_cts = store.cold.slider_per_sector.w_d_wp_1.cts;
+    let new_val = (factor_hh * demand_hh + factor_ind * demand_ind + factor_cts * demand_cts) / (demand_hh + demand_ind + demand_cts);
+    $(`#id_w_d_wp_1`).data("ionRangeSlider").update({from:new_val});
+  }
+  if (slider_id === "id_w_v_3" || slider_id === "id_w_v_4" || slider_id === "id_w_v_5") {
+    let factor_hh = $("#id_w_v_3").data("ionRangeSlider").result.from;
+    let factor_ind = $("#id_w_v_4").data("ionRangeSlider").result.from;
+    let factor_cts = $("#id_w_v_5").data("ionRangeSlider").result.from;
+    let demand_hh = store.cold.slider_per_sector.w_d_wp_1.hh;
+    let demand_ind = store.cold.slider_per_sector.w_d_wp_1.ind;
+    let demand_cts = store.cold.slider_per_sector.w_d_wp_1.cts;
+    let new_val = (factor_hh * demand_hh + factor_ind * demand_ind + factor_cts * demand_cts) / (demand_hh + demand_ind + demand_cts);
+    $(`#id_w_v_1`).data("ionRangeSlider").update({from:new_val});
+  }
+  return logMessage(msg);
 }
 
 function closeSidepanel(panelCloseBtn) {
@@ -218,14 +181,30 @@ function showOrHideSidepanelsOnMoreLabelClick(msg, moreLabel) {
     Array.from(panelSliders).forEach(item => item.parentNode.classList.remove(...classes));
     moreLabel.classList.add(...classes);
   }
-
   return logMessage(msg);
 }
 
 function showActivePanelSliderOnPanelSliderChange(msg, data) {
-  Array.from(panelSliders).forEach(item => item.parentNode.classList.remove("active", "active-sidepanel"));
-  const sliderLabel = data.input[0].parentNode;
-  sliderLabel.classList.add("active");
+  const changedSlider = data.input[0];
+  const changedSliderLabel = changedSlider.parentNode;
+
+  const sliderForm = changedSliderLabel.parentNode.parentNode.parentNode;
+  // Check if any sidepanel is open, by checking if any slider has an active element
+  if (sliderForm.getElementsByClassName("c-slider active").length === 0) {
+    return logMessage(msg);
+  }
+  const isActivePanel = changedSliderLabel.classList.contains("active");
+
+  if (!isActivePanel) {
+    Array.from(panelSliders).forEach(item => {
+      const itemPanel = item.parentNode;
+      if (itemPanel !== changedSliderLabel) {
+        itemPanel.classList.remove("active", "active-sidepanel");
+      }
+    });
+
+    changedSliderLabel.classList.add("active", "active-sidepanel");
+  }
   return logMessage(msg);
 }
 
@@ -265,39 +244,6 @@ function showPVLayers(msg) {
   return logMessage(msg);
 }
 
-function calculate_slider_value(data) {
-  if (data.input[0].id === "id_s_v_3" || data.input[0].id === "id_s_v_4" || data.input[0].id === "id_s_v_5") {
-    let factor_hh = $("#id_s_v_3").data("ionRangeSlider").result.from;
-    let factor_ind = $("#id_s_v_5").data("ionRangeSlider").result.from;
-    let factor_cts = $("#id_s_v_4").data("ionRangeSlider").result.from;
-    let demand_hh = store.cold.slider_per_sector.s_v_1.hh;
-    let demand_ind = store.cold.slider_per_sector.s_v_1.ind;
-    let demand_cts = store.cold.slider_per_sector.s_v_1.cts;
-    let new_val = (factor_hh * demand_hh + factor_ind * demand_ind + factor_cts * demand_cts) / (demand_hh + demand_ind + demand_cts);
-    $(`#id_s_v_1`).data("ionRangeSlider").update({from:new_val});
-  }
-  if (data.input[0].id === "id_w_d_wp_3" || data.input[0].id === "id_w_d_wp_4" || data.input[0].id === "id_w_d_wp_5") {
-    let factor_hh = $("#id_w_d_wp_3").data("ionRangeSlider").result.from;
-    let factor_ind = $("#id_w_d_wp_5").data("ionRangeSlider").result.from;
-    let factor_cts = $("#id_w_d_wp_4").data("ionRangeSlider").result.from;
-    let demand_hh = store.cold.slider_per_sector.w_d_wp_1.hh;
-    let demand_ind = store.cold.slider_per_sector.w_d_wp_1.ind;
-    let demand_cts = store.cold.slider_per_sector.w_d_wp_1.cts;
-    let new_val = (factor_hh * demand_hh + factor_ind * demand_ind + factor_cts * demand_cts) / (demand_hh + demand_ind + demand_cts);
-    $(`#id_w_d_wp_1`).data("ionRangeSlider").update({from:new_val});
-  }
-  if (data.input[0].id === "id_w_v_3" || data.input[0].id === "id_w_v_4" || data.input[0].id === "id_w_v_5") {
-    let factor_hh = $("#id_w_v_3").data("ionRangeSlider").result.from;
-    let factor_ind = $("#id_w_v_5").data("ionRangeSlider").result.from;
-    let factor_cts = $("#id_w_v_4").data("ionRangeSlider").result.from;
-    let demand_hh = store.cold.slider_per_sector.w_d_wp_1.hh;
-    let demand_ind = store.cold.slider_per_sector.w_d_wp_1.ind;
-    let demand_cts = store.cold.slider_per_sector.w_d_wp_1.cts;
-    let new_val = (factor_hh * demand_hh + factor_ind * demand_ind + factor_cts * demand_cts) / (demand_hh + demand_ind + demand_cts);
-    $(`#id_w_v_1`).data("ionRangeSlider").update({from:new_val});
-  }
-}
-
 function calculate_max_wind() {
   let slider_one = $("#id_s_w_5_1").data("ionRangeSlider").result.from / 100;
   let slider_two = $("#id_s_w_5_2").data("ionRangeSlider").result.from / 100;
@@ -316,83 +262,6 @@ function calculate_max_pv_d() {
   let slider = $("#id_s_pv_d_3").data("ionRangeSlider").result.from / 100;
   let new_max = Math.round(slider * store.cold.slider_max.s_pv_d_3);
   $(`#id_s_pv_d_1`).data("ionRangeSlider").update({max:new_max});
-}
-
-function toggleFormFields(formfield_id) {
-  if (formfield_id === "id_s_w_3") {
-    if (document.getElementById("id_s_w_4").checked === false && document.getElementById("id_s_w_5").checked === false) {
-      document.getElementById("id_s_w_3").checked = true;
-    }
-    else {
-      document.getElementById("id_s_w_4").checked = false;
-      document.getElementById("id_s_w_4_1").disabled = true;
-      document.getElementById("id_s_w_4_2").disabled = true;
-      document.getElementById("id_s_w_5").checked = false;
-      $(`#id_s_w_5_1`).data("ionRangeSlider").update({block:true});
-      $(`#id_s_w_5_2`).data("ionRangeSlider").update({block:true});
-      $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_3)});
-    }
-  }
-  if (formfield_id === "id_s_w_4") {
-    if (document.getElementById("id_s_w_3").checked === false && document.getElementById("id_s_w_5").checked === false) {
-      document.getElementById("id_s_w_4").checked = true;
-    }
-    else {
-      if (document.getElementById("id_s_w_4_1").checked === true && document.getElementById("id_s_w_4_2").checked === true) {
-        $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_1) + Math.round(store.cold.slider_max.s_w_4_2)});
-      }
-      if (document.getElementById("id_s_w_4_1").checked === false && document.getElementById("id_s_w_4_2").checked === true) {
-        $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_2)});
-      }
-      if (document.getElementById("id_s_w_4_1").checked === true && document.getElementById("id_s_w_4_2").checked === false) {
-        $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_1)});
-      }
-      document.getElementById("id_s_w_3").checked = false;
-      document.getElementById("id_s_w_4_1").disabled = false;
-      document.getElementById("id_s_w_4_2").disabled = false;
-      document.getElementById("id_s_w_5").checked = false;
-      $(`#id_s_w_5_1`).data("ionRangeSlider").update({block: true});
-      $(`#id_s_w_5_2`).data("ionRangeSlider").update({block: true});
-    }
-  }
-  if (formfield_id === "id_s_w_4_1") {
-    if (document.getElementById("id_s_w_4_2").checked === false) {
-      document.getElementById("id_s_w_4_1").checked = true;
-      $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_1)});
-    }
-    if (document.getElementById("id_s_w_4_1").checked === true && document.getElementById("id_s_w_4_2").checked === true) {
-      $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_1) + Math.round(store.cold.slider_max.s_w_4_2)});
-    }
-    if (document.getElementById("id_s_w_4_1").checked === false && document.getElementById("id_s_w_4_2").checked === true) {
-      $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_2)});
-    }
-  }
-  if (formfield_id === "id_s_w_4_2") {
-    if (document.getElementById("id_s_w_4_1").checked === false) {
-      document.getElementById("id_s_w_4_2").checked = true;
-      $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_2)});
-    }
-    if (document.getElementById("id_s_w_4_1").checked === true && document.getElementById("id_s_w_4_2").checked === true) {
-      $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_1) + Math.round(store.cold.slider_max.s_w_4_2)});
-    }
-    if (document.getElementById("id_s_w_4_1").checked === true && document.getElementById("id_s_w_4_2").checked === false) {
-      $(`#id_s_w_1`).data("ionRangeSlider").update({max:Math.round(store.cold.slider_max.s_w_4_1)});
-    }
-  }
-  if (formfield_id === "id_s_w_5") {
-    if (document.getElementById("id_s_w_3").checked === false && document.getElementById("id_s_w_4").checked === false) {
-      document.getElementById("id_s_w_5").checked = true;
-    }
-    else {
-      document.getElementById("id_s_w_3").checked = false;
-      document.getElementById("id_s_w_4").checked = false;
-      document.getElementById("id_s_w_4_1").disabled = true;
-      document.getElementById("id_s_w_4_2").disabled = true;
-      $(`#id_s_w_5_1`).data("ionRangeSlider").update({block:false});
-      $(`#id_s_w_5_2`).data("ionRangeSlider").update({block:false});
-      calculate_max_wind();
-    }
-  }
 }
 
 function showWindLayers(msg) {